refactor(my-roster): :art: refactor logic and files to improve structure
