+++ app/components/RosterHeader.tsx
@@ -1,4 +1,4 @@
-import { memo, useLayoutEffect, useMemo, useState } from "react"
+import { memo, useMemo } from "react"
 import { useNavigation, useRoute, RouteProp } from "@react-navigation/native"
 import { NativeStackNavigationProp } from "@react-navigation/native-stack"
 import { useSafeAreaInsets } from "react-native-safe-area-context"
