+++ app/screens/DashboardRosterScreen.tsx
@@ -7,7 +7,6 @@ import { Text } from "@/components/Text"
 import { isRTL } from "@/i18n"
 import { DashboardTabScreenProps } from "@/navigators/DashboardNavigator"
 import { useAppTheme } from "@/theme/context"
-import { $styles } from "@/theme/styles"
 import type { ThemedStyle } from "@/theme/types"
 import { openLinkInBrowser } from "@/utils/openLinkInBrowser"
 
@@ -16,7 +15,7 @@ export const DashboardRosterScreen: FC<DashboardTabScreenProps<"DashboardRoster"
     const { themed } = useAppTheme()
 
     return (
-      <Screen preset="scroll" contentContainerStyle={$styles.container}>
+      <Screen preset="scroll">
         <Text preset="heading" tx="dashboardRosterScreen:title" style={themed($title)} />
         <Text tx="dashboardRosterScreen:externalLinkText" style={themed($description)} />
